requirements.txt
django>=4.0.0,<4.2.2
djangorestframework
pyyaml
requests
django-cors-headers